Fixed multiple build creations and started resources
Added a coordinate check to make sure objects are only built at a
location once.
Implemented resources into settlements and Players. Started a README for
things that will eventually need to be in the README. Still need to
create resource cards to be displayed along the bottom of the screen for
each Player.